Run job for JWT expiration notices every 4 hours
This change for a better coverage than when run just once a day (it
was the previous behaviour). Edge case example if we run the job only once a
day: let's say a JWT expires in 25 hours when the CRON executes, the
email alert notifying there is only one day left will not be sent,
it will be sent the day after and 1 hour only before expiration.