Refactor user creation
Contacts were previously created at the same time as the user creation.
Now that they are related to a JwtApiEntreprise user creation as to be
reworked:
* user creation does not create related contacts anymore so just create
  a user account with `email` and `context` fields.
* drop account creation email notices to the previously related contacts

The spec file has also be refactored for better clarity.

No need to test the `note` attribute on user creation since the field is
never provided. This is especially true when the user creation is
automated from the signup form validation (at least for now). We use the
ActiveRecord RSpec helper to ensure the note defaults to an empty
string.