Update user details payload: return tokens' data as with the key itself
Before, only the "secret key" of the tokens were sent in response of a
user details request: since tokens are JWT the client could decode the
base64 payload and extract the tokens details.

Now the tokens details are sent besides the secret key itself. This is
more generic and could be used with non JWT tokens. Therefore it allows
us to send a different "subject" value in the payload (for display
purpose) than the one inside the token payload.

See https://3.basecamp.com/4318089/buckets/14298340/messages/2343969904
for more details about the "subject" trick.