Add some shared contexts / blocks for shared behaviour
Shared examples doesn't work with RSwag, have to use a module with
extend to use them in RSpec DSL.

There's a big limitation : it can only have 1 reponses block with the
same value. Check https://github.com/rswag/rswag/pull/305 for more
information.

It can be a huge no-go (or use the fork ..) for errors like 502 which
bundles a lot of examples, but it can be solved with a lot of examples
in payload.

Moreover, I think we can remove controllers tests, it's redundant.