Refactor mailer for satisfaction survey
Simplify parameters by passing the JWT record only.

This change prevents a false positive that was occuring in the view
specs. The mocked assignment of `:jwt_authorization_request_id` was
hidding its absence in the view layer. Indeed, After refactoring the
mailer and removing the variable instance `@jwt_authorization_request_id`
the specs was still passing (and should have not!).

Be careful with assignments in views specs.