refactor(API::V3): rename into `API::V3AndMore`
For Siade v3, instead of a global API version, we would better like to be able to manage versions per resource. So starting from API v3, each URI should be able to have a dedicated number version that is able to be incremented without having to duplicate its controller. To match this need, the namespace v3 can be renamed into _v3 and more_, meaning v3, and maybe v4, v5, etc (per URI).