DecryptPDF classes no longer fails when qpdf returns warnings
According to the documentation (from qpdf 9.1.1, which is the version
installed on our servers)

```
Ordinarily, qpdf exits with a status of 0 on success or a status of 2
if any errors occurred.  If there were warnings but not errors, qpdf
exits with a status of 3. If warnings would have been issued but --no-warn
was given, an exit status of 3 is still used.
```

The sample PDF has been downloaded from qpdf repository:
https://github.com/qpdf/qpdf/blob/main/qpdf/qtest/qpdf/bad14.pdf

Check https://github.com/qpdf/qpdf/commit/758e3e38f5aceea2502179c7c4840911337f8a3f for more
context.

For the context: there's a new option ( `--warning-exit-0` ) on qpdf >= 10.0.2, but not
available on Ubuntu 20.04 (only after 21.04)

Closes https://gitlab.com/etalab/api-entreprise/siade/-/issues/167