ADEME RGE: Prepare for a new nomenclature
A new nomenclature will be shipped soon which almost reverts to the
previous one, with a tiny difference for 404: it is no longer an empty
array but an array with an entity which has a null id.

This commit handles both actual nomenclature and future nomenclature.

When it'll be shipped, we'll have to remove case described by cassette
`ademe/rge/with_valid_siret.yml` which is a hash (no longer the case in
future nomenclature).

Closes https://gitlab.com/etalab/api-entreprise/siade/-/issues/192