Harden CI by testing that all constants can be loaded through Zeitwerk
Constant autoloading does not work the same way between development/test
and production. The release releases/2022-02-01--17-59 was in a green
state for the CI, but there was a constant loading problem which was not
detected by the CI.

This commit runs a dummy command in the rails production env to check
if the app boots.

There is a default value defined for credentials, which does not change
anything in dev or production: indeed, without production credentials in
the CI, `Siade.credentials` is a nil value. With the {} default value,
this the app does not boot.