FIX Passenger issue with strscan
The commit 162ec3df727f20ac299bb9d0d01e9134b51ced2a introduces a bug on
Passenger:

```
App 2590953 output: Error: The application encountered the following
error: You have already activated strscan 3.0.1, but your Gemfile
requires strscan 3.0.3. Since strscan is a default gem, you can either
remove your dependency on it or try updating to a newer version of
bundler that supports strscan as a default gem. Gem::LoadError)()
```

A (temporary) workaround is to fix strscan to 3.0.1.

The issue seems to be this one:
https://github.com/phusion/passenger/issues/2409

According to these issues, the best workaround is to use
`PassengerPreloadBundler` in order to load right versions of gems.

Ref https://github.com/phusion/passenger/issues/2410#issuecomment-1011906640