Harden Resource implementation
Old implementation used to create methods on the class itself, which is
not really accurate (you can have multiple resource create with a lot of
different data): methods should be accessible only for a specific
object.

A valid approach is to use method_missing which fetch from @data
directly.

Moreover, when you write in cache an instance of Resource and tries to
load it in another ruby thread, old implementation did not enable our
loaded value from marshal to have methods, which causes undefined method
error on production.

Closes https://sentry.data.gouv.fr/organizations/sentry/issues/1049/