Matomo: disable it in production
Seems to block the page loading on page visit (turbo visit) with the
following message (on chrome):

> A parser-blocking, cross site (i.e. different eTLD+1) script,
https://stats.data.gouv.fr/piwik.js, is invoked via document.write. The
network request for this script MAY be blocked by the browser in this or
a future page load due to poor network connectivity. If blocked in this
page load, it will be confirmed in a subsequent console message. See
https://www.chromestatus.com/feature/5718547946799104 for more details.