MI::Associations::ValidateResponse harden 404
The data provider aggregate multiple sources, and can return data with
valid siren/siret which are not associations.

When there is no RNA id and it's not from Alsace Moselle,
we assume it's not an association. Indeed,
since 2009 every new association has a RNA id, and old association get a
RNA id when they add/modify data about their entity, which is seems to
be the case when they add a new document, which is legally required
every year.

For Alsace Moselle, entity does not have RNA.

Source:

* https://www.associations.gouv.fr/nature-des-informations-retournees-10992.html

More info:

* https://associations.gouv.fr/le-rna-repertoire-national-des-associations.html
* https://www.data.gouv.fr/fr/datasets/repertoire-national-des-associations/