yes_jwt helper uses authorizations file
Make it dynamic!

We change TokenFactory to accepts uid, which was fixed with the previous
implementation and required for some tests for DGFIP, which calls
endpoints with a user_id (otherwise VCR does not work).

Clean useless tests (which failed with new jwt) for ACOSS v2

Thanks to this refactor, we no longer have to skip authorization tests
within anonymous controller, which is nice.