Harden cache_key to avoid conflict
Before we have to defined an explicit cache key. It was a choose in
order to be aware that there is cache on a specific endpoint. However,
it did not take into account that a new resource can be build thanks to
api_version for API Entreprise.

Because of it it raises an issue for URSSAF v4 (check issue below).

In order to keep it safe, this commit refactor how we build cache key
for both API Entreprise v3+ and API Particulier (API Entreprise v2 won't
change, so no iteration here).

For API Entreprise, the request path is enough because important params
are within URIs (and query params are only for tracking purpose). It was
not the case on v2 (for example diffusible param for INSEE), but with
the new paradigm it won't be the case.

For API Particulier, request path does not change between calls,
everything is passed in query params, so we used organizer_params method
(which is not the best (for example.. user_id), but good enough for the moment).
There is a specific cache key for PoleEmploi here.

Solves https://github.com/etalab/siade/issues/623
More details: of the original bug: https://github.com/etalab/siade/issues/623#issuecomment-1463891302