DGFIP::AttestationFiscale::MakeRequest timeout to 15
This is a temporary fix in order to mitigate the current problem with
the pdf generation.

The provider does not know when it will be fixed, but after some tests,
it seems to have a generation time between 10s and 15s.

Related https://status.entreprise.api.gouv.fr/clf2asepy57616blobh2jqu7fl