HOTFIX APIEntreprise::V2::DocumentsAssociationsController
In case of non-200 error on a v3 retriever with this cache strategy, retrieved
instance of `Interactor::Context` returns true for `success?`.

Don't know why, so we patch with a check on bundled_data

It's v2, so don't care to migrate this particular endpoint to the v3
philosophy, which implies to rewrite a lot of code.

Moreover, it's 500 error party right now in production: https://errors.data.gouv.fr/organizations/sentry/issues/1636