INSEE::Etablissement: fix when siret is a duplicate
Old behaviour tried to lookup siren from the location path, and called
the search with siege. This is no longer the case: location has directly
the siege social's siret.

On v2 it seems to be a different implementation, don't care because of
deprecation in a few months and no error reported yet.

Closes https://github.com/etalab/siade/issues/703
Closes https://errors.data.gouv.fr/organizations/sentry/issues/336/