API Particulier: track FranceConnect client attributes in logstash
Because of logstasher architecture,
`LogStasher.add_custom_fields_to_request_context` is called before any
processed action within controller.

We have to refactor and lazy assigned the FranceConnect call on a
FranceConnectable controller in order to extract client id/name for
tracking purpose (and avoid multiple calls).
It's not ideal: the refactor within the concern is not straightforward,
but can't find something better for now.

Ref https://linear.app/pole-api/issue/API-980/loguer-le-sub-franceconnect#comment-0c797569