Bugfix EuropeanCommission::VIES::MakeRequest eof reading
Strange issue where the connection seems to be closed before the end of
the TCP closed, seems to be an issue with TLS from the remote server
(not sure).

The TLS issue has been identified thanks to following curl command:

```
curl https://ec.europa.eu/taxation_customs/vies/rest-api/ms/FR/vat/16418166096 --verbose
```

Which renders this (end of command)

```
    "matchCity" : 3,
    "matchCompanyType" : 3
  }
* Closing connection 0
* TLSv1.2 (OUT), TLS header, Supplemental data (23):
* TLSv1.3 (OUT), TLS alert, decode error (562):
}
```

The `decode error` seems to be the issue, there's some thoughts on
the internet about this.

Because of the nature of the unfinished http connection we have to build
a fake http ok response for this specific error.

This is a temporary fix, add some tracking and we should check if it's
solved some day 🤷