Harden tracking error for Sentry
Can't find the root cause of the non-tracking behaviour of this, but
`event` can be another thing than a request, so this block can raise a
NoMethodError (and it's rescued within sentry gem, can be anywhere
there's a lot of rescue with some logging).

Tested directly in production, it works.