Harden tracking of encrypted params for unknown provider errors
Previously without this fix, Sentry calls under the hood `dup`,
which raise the error `TypeError: allocator undefined for GPGME::Data`.

Thanks to `.to_s` `.dup` works.

Can't reproduce the error within tests because of Sentry being disabled:
if the dsn url is not valid no event is sent, and we can't reproduce the
`dup` issue.