Bugfix Rack::Attack.throttled_responder return
According to the documentation, the third argument should be an array of
string, not a string.

Ref https://github.com/rack/rack-attack#customizing-responses
Ref https://www.rubydoc.info/github/rack/rack/file/SPEC.rdoc

Documentation specifies:

```
The Body
The Body is typically an Array of String instances, an enumerable that
yields String instances, a Proc instance, or a File-like object.

The Body must respond to each or call. It may optionally respond to to_path
or to_ary. A Body that responds to each is considered to be an Enumerable Body.
A Body that responds to call is considered to be a Streaming Body.
```