Migrate RedisService from singleton to single class
Prepare to handle all of connection errors within this class, in order
to be more resilient.

We have to drop the singleton in order to keep trying to connect to
redis in case of temporary error.

Moreover redis is now only accessible within the class, in order to keep
it private