Introduce PumaHooks#on_booted
We need to perform an update on a remote API, but we only want to do it
once: that's why we use the `on_booted` hook of Puma.

Because of the override of the puma config by the deployment script in production, we
introduce a `PumaHooks` service in the remote template with the
following config:

```
on_booted do
  if Kernel.const_defined?("PumaHooks") && PumaHooks.new.respond_to?(:on_booted)
    PumaHooks.new.on_booted
  end
end
```

It's safe and can be override per apps.

Ref https://github.com/etalab/very_ansible/pull/457