DatapassWebhook::CreateOrProlongToken sanitized open_data scopes
This prefix enables us to make disjonction on view for DataPass v1/v2.

On DataPass v2 scopes are sent with open_data_ prefix, but not in v1
(thanks to https://github.com/betagouv/datapass/commit/c909e92ea4407d9e0f424011f70e129a493508f7)

This commit sanitizies these entries the same way as DataPass.

It was previously a "hack" on DataPass v1, better to be on this side