Fix error on empty FranceConnect preferred_username
FranceConnect does not return a "preferred_username" key when the data
is empty.
This commit fixes an issue occurring in this case as we considered the
key would always be there: we are using the response body to instanciate
a Resource object which gives access to the data via "method_missing".
When a key is not present it raises as expected.

The workaround in more details: we normalize the response body by
setting the "preferred_username" key to nil when it's not present.
Effectively making the Resource instance work as expected.