Sentry Error - Date naissance was split even though it is null
Sentry issue : https://errors.data.gouv.fr/organizations/sentry/issues/147071/?alert_rule_id=6&alert_timestamp=1728549995293&alert_type=email&environment=production&project=8&referrer=alert_email

This occurs to match the v3 format within the v2.
This is not optimal as this occurs before the parameter validation.
This is not a big deal though as this is suppose to be deleted once the
migration is over