ProlongTokenWizard properly handles draft and submission event
Deleted the listener to the reopen event. When the reopen event, it
means that the authorization_request is a draft and it should not update
the prolong_token_wizard to its "updates_requested" state.

On a reopening, with a submit event, if there is a prolong_token_wizard
expecting for update (ig. its status is "requires_update") then we
should update its status to "updates_requested".