FranceConnect - Removing *_identite scope automatically
When using france connect and api part v3 or more, we remove from the
scopes the *_identite scope allowed for the user.

This is because if the user want to be provided with the user identity
data it must go through france connect.

This behavior is only for api part v3 or more to prevent regression in
api v2.

Those check should be deleted once api part v2 is dead.