Render 409 conflict on multiple same calls
Some users send N times the same request within the same second, which
can lead to a congestion for some providers. This is a way to preserve
them and keep our system and them resilient.

Note for the test: try several things, did not work. I gave up and
override within this test the behaviour of the clean method.
imo it's good enough: code is simple, we put an auto clean after 1
minute too.

Tried to use threads in several ways, unbound methods.. none of them
worked.