Enable Access to API PART v3 swagger
Url is not accessible unless you type it in your browser for now.
This is intended as we can share our url to specific people and continue
our review.
This will also allow syntax checking on the siade_staging_data
repository.

This commit can be reverted once the v3 will be the main swagger on the
siade repository. This has been specified in the dedicated issue on
linear.