QFv2 - Handled specific RNCPS error
This enhances the error message returned by the QFv2 API - both on api
part v2 and v3.

The errorCode provided were given by the api provider. It is very
probable that they will add a specific header code when the error comes
from the RNCPS in the future, and then it should be handled the same way
as the MSA / CNAV specific error.