CNAV - Enhanced validate response to forward 400 from provider
We sometimes have errors from the provider that should not occur. We
can't behave exactly the way they do as they have a very smart system
that make checks that would be too complicated to implement on our side.

BUT, we should not forward a 500 or 503 error when this occurs and let
the user know that there are some issues with his provided params