API PART V3] - Participation familial PSU - Added routes and controller
Nothing special here.

I am not really happy with the tests file the way they are. They induce
heavy reworked once the api is going live and making an API in
"prochainement" should not create work for the live version.

This will be handled in another PR after discussion with the tech team
to enhance the generators and to find a better way to mock behavior
particularly with the france connect modality