API PART V3 - Added 404 file in all operation id
The given params are not exact, but we need to have a 404 file in order
to have a response in case of bad provided params.

Also, if there is no 404 file, then the siade part does not log properly
what has been provided and it's hard to debug.

This should lead to a fix on the siade part to prevent this from
happening