Enhancing CNAV's 404 error management
Explanation of the behavior since it can be hard to understand.

When making a call to the CNAV API we first go through :

- SNGI for user identification
- RNCPS for user attachment to a regime

We can detect wich "sub provider" the error comes from based on
errorCode returned in the response

Then we have the error coming from each regime that are possible.
- Complementaire sante solidaire - RNCPS
- All other CNAV APIs - CNAF or MSA

This is determined from a specific header value as all of the regime
returns the same errors

I could have narrowed and specialised each of the sub validate response
and isolate the CNAF / MSA / RNCPS part, but this has no point since
only the matching API can provide a matching regime.

The key part to understand this is that the RNCPS is involved as an
attachment provider for ALL APIs, and also as a regime only for
complementaire sante solidaire.

In case there is an error in the first two steps, then it's a specific
provider error.
Otherwise, it's a regime related error.

Coming in next commit, the update of all the documentation and tests.