Fixing Effectifs annuels & mensuels payload
Staging was never called and the response was always the `mock_from_api`
because of the forwarded parameter :nature that was forwarded.

Since the 200 response from the staging and the mocked data from the API
are the same we never saw the difference.

All staging request were returning a 200 with the same response