Recipient is a context params
From now on, the recipient is part of the context instead of being part
of a param (which is in the context).

This also provides some additional logic to the recipient
identification. If provided through the api then we pick this one, else
we look for the recipient of the corresponding authorization_request
(which should only happen in API PART v2 AND not with france connect).
If none of these exist (this case should not happen but is a failsafe),
we provide the dinum siret.

In order to prevent regression, the recipient initially provided as a
param within the context have not been removed.