Feature: Make prochainement mocked_params work
The current system works well as long as we don't need
a mocking_params method in the MakeRequest for Staging
to work.

This MakeRequest is added after prochainement mode, which means
in Prochainement mode the staging will never work out of the box if we need
to change the mocking_params.

Now we add MakeRequest before and make it inherit from MockedInteractor
so it should work out of the box.