Normalize how we use recipient across controllers
`recipient` is no longer pass as a param argument to retriever, it's
already passed at the root level of the retriever (like `operation_id`)

We change organizers interface which use recipient by validate the root
level instead of the param.

⚠️ We no longer fallback to DINUM siret for recipient: this fallback
exists for 2 cases:

1. FranceConnect calls without recipient: no one does this in v2 ;
2. Token with authorization requests without siret: there is some cases
   for some API Entreprise authorization requests, only 1 (old) for API
   Particulier

This should not be a breaking change.

Moreover, we are tracking this behaviour

Ref https://linear.app/pole-api/issue/API-5878