Fix: Sentry errors now logging extra_context
We have some Sentry errors not logging extra_context
such as http_body (example:
https://errors.data.gouv.fr/organizations/sentry/issues/268883/?alert_rule_id=6&alert_timestamp=1762462449139&alert_type=email&environment=production&notification_uuid=5928fc9b-5ec3-4300-a60b-808b1f74d1b1&project=8&referrer=alert_email
)

After investigating for a while, the only possible
explanation is that the error actually comes from
ProviderUnknownError in AbstractGetToken,
which doesn't log the extra context.

It is possible that our failing INSEE request are
actually in the authentication part,
not the response validation part.

I propose to merge this fix and see if this solves the issue.